+++ CHANGES.rst
@@ -1,3 +1,33 @@
+Changes in 1.4.4 (2021-06-30)
+=================================================
+
+✨ Features
+ * 
+
+🙌 Improvements
+ * DesignKit: Add Fonts (#4356).
+ * VoIP: Implement audio output router menu in call screen.
+
+🐛 Bugfix
+ * SSO: Handle login callback URL with HTML entities (#4129).
+ * Share extension: Fix theme in dark mode (#4486).
+ * Theme: Fix authentication activity indicator colour when using a dark theme (#4485).
+
+⚠️ API Changes
+ * 
+
+🗣 Translations
+ * 
+    
+🧱 Build
+ * 
+
+Others
+ * 
+
+Improvements:
+ * Upgrade MatrixKit version ([v0.15.3](https://github.com/matrix-org/matrix-ios-kit/releases/tag/v0.15.3)).
+
 Changes in 1.4.3 (2021-06-24)
 =================================================